+++ index.js
@@ -1,32 +1,23 @@
 const _ = require('underscore');
-const Server = require('./server');
+const Server = require('./lib/Server');
 
 let lnurl = {
 	bech32: require('./bech32-rules.json'),
 	createServer(options) {
 		return new Server(options);
 	},
-	generateApiKey(options) {
-		return Server.prototype.generateApiKey(options);
-	},
+	Server,
 };
 
-// Extend the top-level public interface with a few lib methods.
-_.each([
+_.extend(lnurl, _.pick(require('./lib'), [
 	'createAuthorizationSignature',
 	'createSignedUrl',
 	'encode',
 	'decode',
+	'generateApiKey',
 	'generateRandomLinkingKey',
+	'LightningBackend',
 	'verifyAuthorizationSignature',
-], function(method) {
-	lnurl[method] = require(`./lib/${method}`);
-});
-
-// Expose the server prototype.
-lnurl.Server = Server;
-
-// Expose the Lightning backend prototype - for creating custom backends.
-lnurl.LightningBackend = require('./server/LightningBackend');
+]))
 
 module.exports = lnurl;