+++ test/backends/custom.js
@@ -12,27 +12,19 @@ class Backend extends LightningBackend {
 	}
 
 	getNodeUri() {
-		return new Promise((resolve, reject) => {
-			resolve(this.options.nodeUri);
-		});
+		return Promise.resolve(this.options.nodeUri);
 	}
 
 	openChannel(remoteId, localAmt, pushAmt, makePrivate) {
-		return new Promise((resolve, reject) => {
-			resolve();
-		});
+		return Promise.resolve();
 	}
 
 	payInvoice(invoice) {
-		return new Promise((resolve, reject) => {
-			resolve();
-		});
+		return Promise.resolve();
 	}
 
 	addInvoice(amount, extra) {
-		return new Promise((resolve, reject) => {
-			resolve();
-		});
+		return Promise.resolve();
 	}
 }